requirements.txt
matplotlib==3.8.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.1.4
pyserial==3.5
python-dotenv==1.0.0
